backend/AI_Prompt_Engine/requirements.txt
fastapi
uvicorn
python-dotenv
gunicorn
requests
